requirements.txt
astroid==3.0.3
blinker==1.7.0
Cerberus==1.3.5
cfgv==3.4.0
click==8.1.7
dill==0.3.8
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
identify==2.5.33
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.2
pathvalidate==3.2.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
pylint==3.0.3
pypng==0.20220715.0
pytest==8.0.0
python-barcode==0.15.1
PyYAML==6.0.1
qrcode==7.4.2
tomlkit==0.12.3
typing_extensions==4.9.0
virtualenv==20.25.0
Werkzeug==3.0.1
